+++ spec/spec_helper.rb
@@ -2,6 +2,7 @@ $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..'))
 
 require 'rubygems'
 require 'rspec'
+require 'timecop'
 
 require 'slack-ruby-client'